Various fixes.
"noticing an up to 10s increase in application start times"

The next sentence goes on to disagree with this assessment - I think
it's better to just be generic about a "performance regression".

"container spec, and then passes that to
[runc](https://github.com/opencontainers/runc), our container runtime,
which creates and runs a container based on that spec. Before Guardian
calls runc, it asks [GrootFS](https://github.com/cloudfoundry/grootfs)"

Reads like we own runc

"A Pea is our name for a sidecar container. (Docker groups their
sidecars into a Pod, as in a pod of whales. We are Garden so our
sidecars are peas... in a pod."

Pods are a kubernetes concept rather than a docker one.